sql/parser: sanity check builtin functions for length overflow
On 32-bit systems there is a cast from 64-bit int to a 32-bit int
which could cause data correctness issues on large values. Check
for larger values before casting to an int. Also change counter
fields from ints to int64s to be explicit about size.

Updates #8179

+++ sql/parser/aggregate_builtins.go
@@ -181,7 +181,7 @@ func (a *identAggregate) Result() Datum {
 
 type avgAggregate struct {
 	agg   AggregateFunc
-	count int
+	count int64
 }
 
 func newIntAvgAggregate() AggregateFunc {
@@ -273,7 +273,7 @@ func (a *boolOrAggregate) Result() Datum {
 }
 
 type countAggregate struct {
-	count int
+	count int64
 }
 
 func newCountAggregate() AggregateFunc {
@@ -496,7 +496,7 @@ func (a *intVarianceAggregate) Result() Datum {
 }
 
 type floatVarianceAggregate struct {
-	count   int
+	count   int64
 	mean    float64
 	sqrDiff float64
 }
